pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intersect-stem-models"
version = "0.1.0"
description = "Export and packaging tooling for open source music stem separation models."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beartype>=0.14",
  "einops>=0.6.1",
  "numpy>=1.23",
  "onnx>=1.17",
  "onnxscript>=0.5",
  "onnxruntime>=1.20",
  "packaging>=24.0",
  "PyYAML>=6.0",
  "requests>=2.31",
  "rotary-embedding-torch>=0.3.5",
  "torch>=2.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/intersect_stem_models"]
